changes layout
